Local Authentication Added Through Passport.js

+++ controllers/mainController.js
@@ -1,18 +1,14 @@
 const Log = require('../models/logModel')
 // require multer or photo service here
 
-const loginPage = (req, res) => {
-    res.render('login')
-}
-
 const logPage = (req, res) =>  {
-    res.render('log')
+    res.render('log', req.user)
 }
 
 const homePage = async (req, res) => {
     try {
-        const logs = await Log.find()
-        res.render('home', {logs: logs})
+        const logs = await Log.find().populate('user')
+        res.render('home', {logs: logs, user: req.user})
     } catch (error) {
         console.log(error)
     }
@@ -26,11 +22,12 @@ const newLog = async (req, res) => {
             workout1: req.body.workout1,
             workout2: req.body.workout2,
             readingNotes: req.body.readingNotes,
-            otherNotes: req.body.otherNotes
+            otherNotes: req.body.otherNotes,
+            user: req.user._id //attach logged in user id to new log
     })
 
     await log.save()
-    res.redirect('home')
+    res.redirect('/')
 
     } catch (error) {
         console.log(error)
@@ -40,7 +37,7 @@ const newLog = async (req, res) => {
 const editPage = async (req, res) => {
     try {
         const log = await Log.findById(req.params.id)
-        res.render('edit', { log: log })
+        res.render('edit', { log: log, user: req.user })
     } catch (error) {
         console.log(error)
     }
@@ -49,8 +46,11 @@ const editPage = async (req, res) => {
 const updateLog = async (req, res) => {
     console.log(req.params.id, req.body)
     try {
+        let log = await Log.findById(req.params.id)
+        if(log.user.equals(req.user._id)) {
         await Log.findByIdAndUpdate(req.params.id, req.body)
-        res.redirect('/home')
+        }
+        res.redirect('/')
     } catch (error) {
         console.log(error)
     }
@@ -60,15 +60,17 @@ const deleteLog = async (req, res) => {
     console.log(req.params.id)
     console.log(req.body)
     try {
+        let log = await Log.findById(req.params.id)
+        if(log.user.equals(req.user._id)) {
         await Log.findByIdAndRemove(req.params.id)
-        res.redirect('/home')
+        }
+        res.redirect('/')
     } catch (error) {
         console.log(error)
     }
 }
 
 module.exports = {
-    loginPage,
     logPage,
     homePage,
     newLog,
